+++ aiapy/tests/test_idl.py
@@ -6,40 +6,9 @@ from pathlib import Path
 import astropy.units as u
 import numpy as np
 import pytest
-from sunpy import log
 
 from aiapy.calibrate import estimate_error
 from aiapy.conftest import CHANNELS
-
-
-def idl_available():
-    try:
-        import hissw
-
-        hissw.Environment().run("")
-        return True
-    except Exception as e:  # NOQA
-        log.warning(e)
-        return False
-
-
-@pytest.fixture(scope="session")
-def idl_environment():
-    if idl_available():
-        import hissw
-
-        return hissw.Environment(
-            ssw_packages=["sdo/aia"],
-            ssw_paths=["aia"],
-        )
-    pytest.skip(
-        "A working IDL installation is not available. You will not be able to run portions of the test suite.",
-    )
-
-
-@pytest.fixture(scope="session")
-def ssw_home(idl_environment):
-    return idl_environment.ssw_home if idl_available() else None
 
 
 @pytest.mark.parametrize(
