+++ docs/getting_started.rst
@@ -4,8 +4,6 @@
 Getting started with ``aiapy``
 ******************************
 
-``aiapy`` is a community-developed, free and open-source library to to convert AIA images from level 1 to level 1.5, compute point spread function deconvolution, and computing the wavelength response functions for the EUV channels.
-
 .. toctree::
     :maxdepth: 1
