+++ docs/index.rst
@@ -4,9 +4,9 @@
 aiapy documentation
 ===================
 
-``aiapy`` is a Python library for analyzing data from the Atmospheric Imaging Assembly (AIA) instrument onboard the Solar Dynamics Observatory (SDO) spacecraft.
+``aiapy`` is a community-developed, free and open-source Python library for analyzing data from the Atmospheric Imaging Assembly (AIA) onboard the Solar Dynamics Observatory (SDO).
 
-It allows for converting AIA images from level 1 to level 1.5, point spread function deconvolution, and computing the wavelength response functions for the EUV channels.
+Capabilities include converting AIA images from level 1 to level 1.5, deconvolving images with the instrument point spread function, and computing the wavelength response functions for the EUV channels.
 
 .. grid:: 1 2 2 2
     :gutter: 2
